app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
alembic==1.12.1
tenacity==8.2.3
redis==5.0.1
celery==5.3.6
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
prometheus-client==0.11.0
pytest==6.2.5
httpx==0.19.0 